requirements.txt
streamlit
google-generativeai
mcp
python-dotenv
nest_asyncio
httpx
sse-starlette

